+++ modules/amongus.py
@@ -39,14 +39,16 @@ class AmongUsCog(commands.Cog, name="Among Us Cog"):
         players = json.load(open(self.db_file))
         players["voting"] = False
         json.dump(players, open(self.db_file, "w"), indent=4)
-        await ctx.message.delete()
+        if not isinstance(ctx.channel, discord.DMChannel) or not isinstance(ctx.channel, discord.GroupChannel):
+            await ctx.message.delete()
 
     @commands.command(name="during_voting", aliases=["dv"])
     async def during_voting(self, ctx):
         players = json.load(open(self.db_file))
         players["voting"] = True
         json.dump(players, open(self.db_file, "w"), indent=4)
-        await ctx.message.delete()
+        if not isinstance(ctx.channel, discord.DMChannel) or not isinstance(ctx.channel, discord.GroupChannel):
+            await ctx.message.delete()
 
     @commands.command(name="start_game", aliases=["sg"])
     async def start_game(self, ctx):
@@ -63,7 +65,8 @@ class AmongUsCog(commands.Cog, name="Among Us Cog"):
                 players["all_players"].append(u.id)
         json.dump(players, open(self.db_file, "w"), indent=4)
         self.during_game.start()
-        await ctx.message.delete()
+        if not isinstance(ctx.channel, discord.DMChannel) or not isinstance(ctx.channel, discord.GroupChannel):
+            await ctx.message.delete()
 
     @commands.command(name="end_game", aliases=["eg"])
     async def end_game(self, ctx):
@@ -72,7 +75,8 @@ class AmongUsCog(commands.Cog, name="Among Us Cog"):
         for u in self.vc.members:
             await u.edit(deafen=False, mute=False)
         self.during_game.cancel()
-        await ctx.message.delete()
+        if not isinstance(ctx.channel, discord.DMChannel) or not isinstance(ctx.channel, discord.GroupChannel):
+            await ctx.message.delete()
 
     @commands.command(name="pdied", aliases=["pd"])
     async def pdied(self, ctx, player: discord.Member):
@@ -80,7 +84,8 @@ class AmongUsCog(commands.Cog, name="Among Us Cog"):
         players["alive"].remove(player.id)
         players["dead"].append(player.id)
         json.dump(players, open(self.db_file, "w"), indent=4)
-        await ctx.message.delete()
+        if not isinstance(ctx.channel, discord.DMChannel) or not isinstance(ctx.channel, discord.GroupChannel):
+            await ctx.message.delete()
 
     async def cog_command_error(self, ctx, error):
         error = getattr(error, 'original', error)
