+++ modules/points.py
@@ -24,7 +24,8 @@ class PointsCog(commands.Cog, name="Points Commands"):
 
     @commands.command(aliases=["hp"])
     async def house_points(self, ctx):
-        await ctx.message.delete()
+        if not isinstance(ctx.channel, discord.DMChannel) or not isinstance(ctx.channel, discord.GroupChannel):
+            await ctx.message.delete()
         data = json.load(open(self.db_file))
         house_emb = discord.Embed(title="House points", colour=0x00adff)
         for house in data["houses"]:
@@ -34,7 +35,8 @@ class PointsCog(commands.Cog, name="Points Commands"):
     @commands.cooldown(1, 180, commands.BucketType.user)
     @commands.command(aliases=["cs"])
     async def cast_spell(self, ctx):
-        await ctx.message.delete()
+        if not isinstance(ctx.channel, discord.DMChannel) or not isinstance(ctx.channel, discord.GroupChannel):
+            await ctx.message.delete()
         data = json.load(open(self.db_file))
         if random.randint(1, 10) >= 5:
             if random.randint(1, 10) >= 5:
@@ -67,7 +69,8 @@ class PointsCog(commands.Cog, name="Points Commands"):
     @commands.cooldown(1, 600, commands.BucketType.user)
     @commands.command()
     async def steal(self, ctx, house_name):
-        await ctx.message.delete()
+        if not isinstance(ctx.channel, discord.DMChannel) or not isinstance(ctx.channel, discord.GroupChannel):
+            await ctx.message.delete()
         data = json.load(open(self.db_file))
         houses_steal_from = data["houses"].copy()
         for house in houses_steal_from:
@@ -105,7 +108,8 @@ class PointsCog(commands.Cog, name="Points Commands"):
 
     @commands.command()
     async def beg(self, ctx):
-        await ctx.message.delete()
+        if not isinstance(ctx.channel, discord.DMChannel) or not isinstance(ctx.channel, discord.GroupChannel):
+            await ctx.message.delete()
         data = json.load(open(self.db_file))
         if ctx.message.author.id in [i["person"] for i in data["timeouts"]]:
             for i in data["timeouts"]:
@@ -157,7 +161,8 @@ class PointsCog(commands.Cog, name="Points Commands"):
     @commands.command()
     async def starvetodeath(self, ctx):
         data = json.load(open(self.db_file))
-        await ctx.message.delete()
+        if not isinstance(ctx.channel, discord.DMChannel) or not isinstance(ctx.channel, discord.GroupChannel):
+            await ctx.message.delete()
         if data["commie_times"] != 5:
             await ctx.send(data["commie_strings"][data["commie_times"]])
             data["commie_times"] += 1
